fix(voice): address PR #5942 review — move VoiceEvent to gateway, add error frames, reorder config check (#5896)
- Move VoiceEvent enum + serde roundtrip tests from zeroclaw-api/src/vad.rs
  to zeroclaw-gateway/src/voice_duplex.rs (RFC #5574 §4.2 compliance)
- Vad trait, VadEvent, NoopVad remain in zeroclaw-api (correct placement)
- handle_voice_event now returns Option<serde_json::Value> error frame
  for server→client events received from client (invalid_event_direction)
- ws.rs: check runtime config before parsing voice events (avoid redundant
  JSON parse on every message when duplex compiled but not enabled)
- Add tests for error frame and no-error paths
