+++ crates/zeroclaw-gateway/src/ws.rs
@@ -333,19 +333,20 @@ async fn handle_socket(
                 // ── Voice duplex event dispatch (gated by feature flag + runtime config) ──
                 #[cfg(feature = "gateway-voice-duplex")]
                 {
-                    if let Some(voice_event) = crate::voice_duplex::try_parse_voice_event(&msg) {
-                        let duplex_enabled = state
-                            .config
-                            .lock()
-                            .channels
-                            .voice_duplex
-                            .as_ref()
-                            .is_some_and(|v| v.enabled);
-                        if duplex_enabled {
-                            crate::voice_duplex::handle_voice_event(voice_event);
+                    let duplex_enabled = state
+                        .config
+                        .lock()
+                        .channels
+                        .voice_duplex
+                        .as_ref()
+                        .is_some_and(|v| v.enabled);
+                    if duplex_enabled {
+                        if let Some(voice_event) = crate::voice_duplex::try_parse_voice_event(&msg) {
+                            if let Some(error_frame) = crate::voice_duplex::handle_voice_event(voice_event) {
+                                let _ = sender.send(Message::Text(error_frame.to_string().into())).await;
+                            }
                             continue;
                         }
-                        // Duplex not enabled at runtime → fall through to unknown type error
                     }
                 }
